GDK/Win32: Remove stacking functions
Just take into account the Z-order when calling SetWindowPos() instead of
explicitly re-stacking the windows.  This will ensure that dialog boxes and
pickers do not somehow fall behind another (gtk)window, and avoids spews of
warnings when we put a parent window of a popover before the popover itself.

This will fix restacking on (re-)activation, as the stacking functions caused
numerous issues in terms of stacking.

Forward port of LRN's patch "GDK W32: transient Z-order rollback" in MR !1769